lpl-slack-bot-test/requirements.txt
requests-html==0.10.0
slack-sdk==3.19.0
python-dotenv==0.21.0
lxml[html_clean]
psycopg2==2.9.1
